requirements.txt
torch>=1.6.0
torchaudio>=0.6.0
numpy
scipy
scikit-learn
tqdm
opencv-python
python_speech_features
torchvision
Pillow==9.0.0